+++ Code/GraphMol/QueryOps.cpp
@@ -10,6 +10,7 @@
 #include "QueryOps.h"
 #include <algorithm>
 #include <RDGeneral/types.h>
+#include <GraphMol/QueryAtom.h>
 
 namespace RDKit {
 
@@ -503,6 +504,13 @@ ATOM_EQUALS_QUERY *makeAtomHasAliphaticHeteroatomNbrsQuery() {
   return res;
 }
 
+ATOM_EQUALS_QUERY *makeAtomNonHydrogenDegreeQuery(int what) {
+  auto *res =
+      makeAtomSimpleQuery<ATOM_EQUALS_QUERY>(what, queryAtomNonHydrogenDegree);
+  res->setDescription("AtomNonHydrogenDegree");
+  return res;
+}
+
 BOND_EQUALS_QUERY *makeBondOrderEqualsQuery(Bond::BondType what) {
   auto *res = new BOND_EQUALS_QUERY;
   res->setVal(what);
@@ -711,4 +719,52 @@ bool isAtomAromatic(const Atom *a) {
   }
   return res;
 }
+
+namespace QueryOps {
+namespace {
+void completeQueryAndChildren(ATOM_EQUALS_QUERY *query, Atom *tgt,
+                              unsigned int magicVal) {
+  PRECONDITION(query, "no query");
+  PRECONDITION(tgt, "no atom");
+  if (static_cast<unsigned int>(query->getVal()) == magicVal) {
+    int tgtVal = query->getDataFunc()(tgt);
+    query->setVal(tgtVal);
+  }
+  for (auto childIt = query->beginChildren(); childIt != query->endChildren();
+       ++childIt) {
+    completeQueryAndChildren((ATOM_EQUALS_QUERY *)(childIt->get()), tgt,
+                             magicVal);
+  }
+}
+}  // namespace
+void completeMolQueries(RWMol *mol, unsigned int magicVal) {
+  PRECONDITION(mol, "bad molecule");
+  for (auto atom : mol->atoms()) {
+    if (atom->hasQuery()) {
+      auto *query = static_cast<ATOM_EQUALS_QUERY *>(atom->getQuery());
+      completeQueryAndChildren(query, atom, magicVal);
+    }
+  }
+}
+
+Atom *replaceAtomWithQueryAtom(RWMol *mol, Atom *atom) {
+  PRECONDITION(mol, "bad molecule");
+  PRECONDITION(atom, "bad atom");
+  if (atom->hasQuery()) {
+    return atom;
+  }
+
+  QueryAtom qa(*atom);
+  unsigned int idx = atom->getIdx();
+
+  if (atom->getFormalCharge() != 0) {
+    qa.expandQuery(makeAtomFormalChargeQuery(atom->getFormalCharge()));
+  }
+  if (atom->hasProp(common_properties::_hasMassQuery)) {
+    qa.expandQuery(makeAtomMassQuery(static_cast<int>(atom->getMass())));
+  }
+  mol->replaceAtom(idx, &qa);
+  return mol->getAtomWithIdx(idx);
+}
+}  // namespace QueryOps
 };  // namespace RDKit